Catch errors and provide warnings for trying to get RNA-seq and dbSNP resources that are not available. Thanks to @brainstorm. Closes #67

+++ cloudbio/biodata/rnaseq.py
@@ -20,13 +20,17 @@ def download_transcripts(genomes, env):
         tx_dir = os.path.join(org_dir, gid, folder_name)
         if not exists(tx_dir):
             with cd(org_dir):
-                _download_annotation_bundle(base_url.format(gid=gid))
+                _download_annotation_bundle(base_url.format(gid=gid), gid)
 
-def _download_annotation_bundle(url):
+def _download_annotation_bundle(url, gid):
     """Download bundle of RNA-seq data from S3 biodata/annotation
     """
     tarball = os.path.basename(url)
     if not exists(tarball):
-        run("wget %s" % url)
-    run("xz -dc %s | tar -xvpf -" % tarball)
-    run("rm -f %s" % tarball)
+        with warn_only():
+            run("wget %s" % url)
+    if exists(tarball):
+        run("xz -dc %s | tar -xvpf -" % tarball)
+        run("rm -f %s" % tarball)
+    else:
+        env.logger.warn("RNA-seq transcripts not available for %s" % gid)
